refactor(core): move types
