+++ packages/core/src/config/robots/index.ts
@@ -1,3 +1,3 @@
-import { Mode, Robots } from "../types";
+import { Mode, Robots } from "../../types";
 
 export default ({ mode }: { mode: Mode }): Robots => "User-agent: *\nDisallow:";
