+++ packages/core/src/config/webpack/devtool.ts
@@ -1,5 +1,5 @@
 import { Configuration } from "webpack";
-import { Target, Mode } from "../types";
+import { Target, Mode } from "../../types";
 
 export default ({
   target,
